Added/Universalized Pages
Added a header to all the pages, now there's a page flow.  Rejoined the
collections.  The database functions all in one collection.  The
differentiation between the pages is that they have a document
identifier that specifies if we're operating in the Spec Library or the
Inventory Library.
There was also some restrictions added to the input.  The input must be
unique for unit classes in Specs and for skus in inventory.
Coming soon:  Keyboard shortcuts, more Spec Library features for
adding/removing various hierarchical elements.