Group Page updated
Groups page now has grid of groups in the db.  The user can select
fields from the fields grid, see which fields are selected in the
selection grid, give the group a name and commit it to the db and see
the group grid update to include their new group.  I also restricted
additions to only allow unique group names.  There also must be a name
provided and the user must select at least one field.  Next step is to
add another grid that displays the fields included in the selected
group.