Group/Product Pages
Group Page has added group removal functionality, the interface was
simplified a bit.  And another grid was added that displays the fields
of the currently selected group.
The product builder page was created.  it lists the available groups.
The user can select from the available groups to define a new product.
A second grid displays all the parameters that are selected from the
selected groups.  Soon to come, there will be a list of completed
products that will also have remove functionality.