Product Page: Product creation now added
New Products can officially be added to the db.  They are also listed
in a grid once they are added.  Next up: removal of products from the
db and grid.  And a grid that displays all the specs of the selected
product and allows the user to set values for each spec.
Specs are saved in each product with an initial value of 0.