Parameter Page makeover
The Parameter page got a nice makeover.  The buttons are now
successfully dynamically added to the grid with each parameter.  Each
row contains a delete button that removes the Parameter from the grid
and db.  There's also a get button for the value and unit of each
parameter in the grid.  The get buttons place text into the
"Definition" textbox to define a computed field.  The format of the
"definition" text is still undecided.  The definition text will then be
parsed and displayed in the "Output" textbox.
The "Remove" button was removed because it is now obsolete due to the
delete buttons in the grid.
The "Insert" button was updated to be labeled more specifically and to
clear the text fields if a submission is successful.  The db
interaction was also updated to include the value and unit fields and
the uniqueness requirement of the parameter name is no longer applied
because uniqueness if implied by the combination of name, value, and
field.