Computed Fields update
The computed fields portion of the parameter page is somewhat working.
It's not the most user friendly experience, so would like to edit it
further.  It currently adds the element to the definition when the user
clicks the get button from the grid.  It separates elements by a
semi-colon.  If the user would like to add text between elements, they
add a semi-colon and click the "Append Text" button.  Then they are
free to add more elements to the definition.  Currently none of this
data gets saved to the database.  All files also received a credit and
use notice at the top of the documentation.