Computed Fields commitable
The computed fields are now commitable to the db.  They are of a
different type than standard fields because they possess different
properties.  The entire definition string is stored along with a name.
A grid was added to the page to display the computed fields that exist
in the db.  It functions similarly to the Parameter grid: it shows a
row number, a name, the definition, a get button to copy the entire
definition to the definition input box, and a delete button to remove
the computed field from the database.

Some restrictions were added to the text fields to correctly require
certain fields depending on whether the checkbox to enter a computed
field is checked.

The submit new parameter button checks if the checkbox is checked and
this is how it determines what time of field to commit to the database.

The definition text fields were widened a bit to accomodate for their
lengthy content.

Next step, the view builder should now be more comprehendable to put
together.  The view builder will allow the user to create and customize
how spec reports are displayed.