Layout Overhaul 1.0
Overhaul on the layout.  Each Page is laid out to have three main
views.  The 'Create New' view, the 'List' view, and the 'Viewer' view.
The create new view is where a new objects members are chosen and/or
created.  The list view is where all the objects that exist are listed,
and the viewer view displays a selected objects members.
Each list view now possesses a delete button for each item in its grid.