Grid Details
Most grids now display types.  The database got an overhaul that
streamlines types.  The view grid on each page is also capable of
editing and storing changes.  The major bug at the moment is ensuring
no loss of computed field definitions when editing fields in the viewer
on the groups and product pages.