v1
Changed the parameter grids so they split vertically instead of
horizontally.  Made the computed input a little bit more short-hand.
Needs to be further optimized.  The groups now reference individual
paramter ids rather than parameter's data.  It stores an array of ids
instead of an array of JSON objects.  The intent is that when one of
those ids change, the group data will change.  I haven't been able to
get the data to display properly yet.

This is a new branch because we're going to try to go head on toward
asynchronous behavior.  The point is to have the grids really updating
live upon the data.  The first attempt will be to create sub-stores of
data and see if we can get the grid to update upon that sub-store on
the fly.  Once we get that implemented, we will implement a global
store of all the parameters to which the substores "subscribe."  So if
any of the global store's data is changed, all sub stores that are
subscribed to that data will update, and further, the associated grids
will update.